feat: Use Bookworm base images for smaller Debian-based image
Replaces Alpine base images with Debian Bookworm in the Dockerfile.

This change leverages Debian Bookworm and Bookworm Slim as base images
for the builder and runner stages respectively, instead of Alpine.

While Alpine is generally smaller, It was not compatible with tailwindcss cli app.
Thus I had to use this version which was compatible with it.

The Dockerfile remains a multi-stage build, ensuring a minimal runner
image by copying only the necessary executable from the builder stage.